+++ packages/sdk/src/get_sdk.ts
@@ -1,7 +1,7 @@
 import SdkType from './sdk';
 
 // TODO(alex): prevent sdk sharing across invocations of the same lambda container
-let sdk: SdkType;
+let sdk: SdkType | null = null;
 
 /** @hidden */
 export default function getSdk(): SdkType {
@@ -11,5 +11,10 @@ export default function getSdk(): SdkType {
         sdk = new Sdk(getAirtableInterface());
     }
 
-    return sdk;
+    return sdk as SdkType;
+}
+
+/** @hidden */
+export function clearSdkForTest() {
+    sdk = null;
 }