+++ packages/sdk/src/types/airtable_interface.ts
@@ -30,6 +30,7 @@ export interface SdkInitData {
     isFullscreen: boolean;
     isFirstRun: boolean;
     intentData: unknown;
+    isUsingNewLookupCellValueFormat?: true | undefined;
 }
 
 /** @hidden */